requirements.txt
# Dépendances Python pour l'application Prix Carburants

# Framework web
Flask==3.0.0
Flask-CORS==4.0.0

# Requêtes HTTP
requests==2.31.0

# Parsing XML
lxml==4.9.3

# Gestion des dates
python-dateutil==2.8.2

# Serveur WSGI pour la production
gunicorn==21.2.0

# Monitoring et logging (optionnel)
python-dotenv==1.0.0

# Sécurité
Werkzeug==3.0.1

# Développement et tests (optionnel)
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
